# internal/handler/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger specification for a Duel Duck back-end.",
        "title": "Duel Duck API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refreshes the user's JWT tokens using the provided refresh token from the Authorization header. Returns a new pair of access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh JWT tokens",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003crefresh_token\u003e",
                        "description": "Bearer refresh token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully refreshed tokens",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "jwt_info": {
                                    "$ref": "#/definitions/auth.TokenPair"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
...
        "/auth/sign-in-google": {
            "post": {
                "security": [
                    {
                        "GoogleAuth": []
                    }
                ],
                "description": "Authenticates a user using Google credentials and returns user data with JWT tokens. The email is validated through Google's authentication system and a new session is created.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in user with Google authentication",
                "parameters": [
                    {
                        "description": "Google sign in request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignInWithGoogle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "daily_reward": {
                                    "type": "integer"
                                },
                                "jwt_info": {
                                    "$ref": "#/definitions/auth.TokenPair"
                                },
                                "user": {
                                    "$ref": "#/definitions/model.User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data or missing email",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid Google token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/auth/sign-in-telegram": {
            "post": {
                "description": "Authenticates a user using Telegram Mini App init data. Validates the init data hash and creates/updates user account. Returns user data and JWT tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in user with Telegram Mini App",
                "parameters": [
                    {
                        "description": "Telegram init data and additional parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignInWithTelegramMiniAppReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "daily_reward": {
                                    "type": "integer"
                                },
                                "jwt_info": {
                                    "type": "object",
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "user": {
                                    "$ref": "#/definitions/model.User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data or failed to parse init data",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Invalid telegram init data hash or login timeout",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/auth/sign-in-telegram-web": {
            "post": {
                "description": "Authenticates a user using previously generated Telegram authentication data. Validates the provided code against stored code for the given Telegram ID and creates/updates user account. Returns user data and JWT tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in user with Telegram Web authentication",
                "parameters": [
                    {
                        "description": "Telegram authentication data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignInWithTelegramWeb"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "daily_reward": {
                                    "type": "integer"
                                },
                                "jwt_info": {
                                    "$ref": "#/definitions/auth.TokenPair"
                                },
                                "user": {
                                    "$ref": "#/definitions/model.User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired authentication code",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperrors.ErrorPublic": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "auth.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "model.SignInWithGoogle": {
            "type": "object",
            "properties": {
                "referrer_token": {
                    "type": "string"
                }
            }
        },
        "model.SignInWithTelegramMiniAppReq": {
            "type": "object",
            "properties": {
                "init_data_raw": {
                    "type": "string"
                },
                "referral_token": {
                    "type": "string"
                }
            }
        },
        "model.SignInWithTelegramWeb": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "bg_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_xp": {
                    "type": "integer"
                },
                "daily_reward_streak": {
                    "type": "integer"
                },
                "discord_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "instagram_url": {
                    "type": "string"
                },
                "is_premium": {
                    "type": "boolean"
                },
                "last_completed_streak": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "public_address": {
                    "type": "string"
                },
                "referral_token": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/mtype.Role"
                },
                "sol_autoswap": {
                    "type": "boolean"
                },
                "telegram_id": {
                    "type": "string"
                },
                "telegram_url": {
                    "type": "string"
                },
                "two_fa": {
                    "type": "boolean"
                },
                "two_fa_secret": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usdc_autoswap": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "x_url": {
                    "type": "string"
                },
                "youtube_url": {
                    "type": "string"
                }
            }
        },
...
        "mtype.Role": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "UserRole",
                "AdminRole"
            ]
        },
...
    }
}